Split the multiple test classes into their own file and fixed up the pwsh assertion
